# chore(consensus): Simplify `IngressPayload`