--- a/rs/types/types/src/batch/ingress.rs
+++ b/rs/types/types/src/batch/ingress.rs
@@ -7,38 +7,82 @@ use crate::{
 use ic_exhaustive_derive::ExhaustiveSet;
 use ic_protobuf::{proxy::ProxyDecodeError, types::v1 as pb};
 use serde::{Deserialize, Serialize};
-use std::{
-    convert::TryFrom,
-    io::{Cursor, Write},
-};
+use std::convert::TryFrom;
+
+#[derive(Clone, Debug, PartialEq, Eq, Hash, Serialize, Deserialize)]
+struct SerializedIngressMessage {
+    ingress_message_id: IngressMessageId,
+    #[serde(with = "serde_bytes")]
+    buffer: Vec<u8>,
+}
 
 /// Payload that contains Ingress messages
 #[derive(Clone, Debug, Default, PartialEq, Eq, Hash, Serialize, Deserialize)]
 #[cfg_attr(test, derive(ExhaustiveSet))]
 pub struct IngressPayload {
-    /// Pairs of MessageId and its serialized byte position in the buffer.
-    id_and_pos: Vec<(IngressMessageId, u64)>,
-    /// All messages are serialized in a single byte buffer, so individual
-    /// deserialization is delayed. This allows faster deserialization of
-    /// IngressPayload when individual message is not needed (e.g. in
-    /// ingress payload deduplication).
-    #[serde(with = "serde_bytes")]
-    buffer: Vec<u8>,
+    ingress_messages: Vec<SerializedIngressMessage>,
 }
 
-impl From<&IngressPayload> for pb::IngressPayload {
-    fn from(ingress_payload: &IngressPayload) -> Self {
+impl From<SerializedIngressMessage> for pb::IngressMessage {
+    fn from(message: SerializedIngressMessage) -> Self {
         Self {
-            id_and_pos: ingress_payload
-                .id_and_pos
-                .iter()
-                .map(|(msg_id, offset)| pb::IngressIdOffset {
-                    expiry: msg_id.expiry().as_nanos_since_unix_epoch(),
-                    message_id: msg_id.message_id.as_bytes().to_vec(),
-                    offset: *offset,
-                })
+            expiry: message
+                .ingress_message_id
+                .expiry()
+                .as_nanos_since_unix_epoch(),
+            message_id: message.ingress_message_id.message_id.as_bytes().to_vec(),
+            ingress_message: message.buffer,
+        }
+    }
+}
+
+impl TryFrom<pb::IngressMessage> for SerializedIngressMessage {
+    type Error = ProxyDecodeError;
+
+    fn try_from(message: pb::IngressMessage) -> Result<Self, Self::Error> {
+        Ok(Self {
+            ingress_message_id: IngressMessageId::new(
+                Time::from_nanos_since_unix_epoch(message.expiry),
+                MessageId::try_from(message.message_id.as_slice())?,
+            ),
+            buffer: message.ingress_message,
+        })
+    }
+}
+
+impl From<SignedIngress> for SerializedIngressMessage {
+    fn from(ingress: SignedIngress) -> Self {
+        Self {
+            ingress_message_id: IngressMessageId::from(&ingress),
+            buffer: ingress.binary().as_ref().to_vec(),
+        }
+    }
+}
+
+impl TryFrom<SerializedIngressMessage> for SignedIngress {
+    type Error = IngressPayloadError;
+
+    fn try_from(ingress_message: SerializedIngressMessage) -> Result<Self, Self::Error> {
+        let ingress = SignedIngress::try_from(SignedRequestBytes::from(ingress_message.buffer))
+            .map_err(|e| IngressPayloadError::DeserializationFailure(e.to_string()))?;
+        let ingress_id = IngressMessageId::from(&ingress);
+        if ingress_message.ingress_message_id == ingress_id {
+            Ok(ingress)
+        } else {
+            // FIXME:
+            Err(IngressPayloadError::MismatchedMessageIdAtIndex(0))
+        }
+    }
+}
+
+impl From<IngressPayload> for pb::IngressPayload {
+    fn from(ingress_payload: IngressPayload) -> Self {
+        Self {
+            ingress_messages: ingress_payload
+                .ingress_messages
+                .into_iter()
+                .map(pb::IngressMessage::from)
                 .collect(),
-            buffer: ingress_payload.buffer.clone(),
         }
     }
 }
@@ -48,20 +92,11 @@ impl TryFrom<pb::IngressPayload> for IngressPayload {
 
     fn try_from(payload: pb::IngressPayload) -> Result<Self, Self::Error> {
         Ok(Self {
-            id_and_pos: payload
-                .id_and_pos
-                .iter()
-                .map(|ingress_offset| {
-                    Ok((
-                        IngressMessageId::new(
-                            Time::from_nanos_since_unix_epoch(ingress_offset.expiry),
-                            MessageId::try_from(ingress_offset.message_id.as_slice())?,
-                        ),
-                        ingress_offset.offset,
-                    ))
-                })
-                .collect::<Result<Vec<_>, Self::Error>>()?,
-            buffer: payload.buffer,
+            ingress_messages: payload
+                .ingress_messages
+                .into_iter()
+                .map(SerializedIngressMessage::try_from)
+                .collect::<Result<_, _>>()?,
         })
     }
 }
@@ -84,20 +119,20 @@ pub enum IngressPayloadError {
 impl IngressPayload {
     /// Return the number of ingress messages contained in this payload
     pub fn message_count(&self) -> usize {
-        self.id_and_pos.len()
+        self.ingress_messages.len()
     }
 
     /// Return all MessageIds in the payload.
     pub fn message_ids(&self) -> Vec<IngressMessageId> {
-        self.id_and_pos
+        self.ingress_messages
             .iter()
-            .map(|(id, _)| id.clone())
+            .map(|ingress_message| ingress_message.ingress_message_id.clone())
             .collect::<Vec<_>>()
     }
 
     /// Return true if the payload is empty.
     pub fn is_empty(&self) -> bool {
-        self.id_and_pos.is_empty()
+        self.ingress_messages.is_empty()
     }
 
     /// Return the ingress message at a given index, which is expected to be
@@ -106,31 +141,19 @@ impl IngressPayload {
         &self,
         index: usize,
     ) -> Result<(IngressMessageId, SignedIngress), IngressPayloadError> {
-        self.id_and_pos
+        self.ingress_messages
             .get(index)
             .ok_or(IngressPayloadError::IndexOutOfBound(index))
-            .and_then(|(id, pos)| {
-                // Return error if pos is out of bound.
-                if *pos > self.buffer.len() as u64 {
-                    Err(IngressPayloadError::IngressPositionOutOfBound(index, *pos))
+            .and_then(|ingress_message| {
+                let ingress = SignedIngress::try_from(SignedRequestBytes::from(
+                    ingress_message.buffer.clone(),
+                ))
+                .map_err(|e| IngressPayloadError::DeserializationFailure(e.to_string()))?;
+                let ingress_id = IngressMessageId::from(&ingress);
+                if ingress_message.ingress_message_id == ingress_id {
+                    Ok((ingress_id, ingress))
                 } else {
-                    let end = {
-                        if index == self.id_and_pos.len() - 1 {
-                            self.buffer.len()
-                        } else {
-                            self.id_and_pos[index + 1].1 as usize
-                        }
-                    };
-                    let ingress = SignedIngress::try_from(SignedRequestBytes::from(Vec::from(
-                        &self.buffer[*pos as usize..end],
-                    )))
-                    .map_err(|e| IngressPayloadError::DeserializationFailure(e.to_string()))?;
-                    let ingress_id = IngressMessageId::from(&ingress);
-                    if *id == ingress_id {
-                        Ok((ingress_id, ingress))
-                    } else {
-                        Err(IngressPayloadError::MismatchedMessageIdAtIndex(index))
-                    }
+                    Err(IngressPayloadError::MismatchedMessageIdAtIndex(index))
                 }
             })
     }
@@ -138,26 +161,20 @@ impl IngressPayload {
 
 impl CountBytes for IngressPayload {
     fn count_bytes(&self) -> usize {
-        self.buffer.len() + self.id_and_pos.len() * EXPECTED_MESSAGE_ID_LENGTH
+        self.ingress_messages
+            .iter()
+            .map(|message| message.buffer.len() + EXPECTED_MESSAGE_ID_LENGTH)
+            .sum()
     }
 }
 
 impl From<Vec<SignedIngress>> for IngressPayload {
     fn from(msgs: Vec<SignedIngress>) -> IngressPayload {
-        let mut buf = Cursor::new(Vec::new());
-        let mut id_and_pos = Vec::new();
-        for ingress in msgs {
-            let id = IngressMessageId::from(&ingress);
-            let pos = buf.position();
-            // This panic will only happen when we run out of memory.
-            buf.write_all(ingress.binary().as_ref())
-                .unwrap_or_else(|err| panic!("SignedIngress serialization error: {:?}", err));
-
-            id_and_pos.push((id, pos));
-        }
-        IngressPayload {
-            id_and_pos,
-            buffer: buf.into_inner(),
+        Self {
+            ingress_messages: msgs
+                .into_iter()
+                .map(SerializedIngressMessage::from)
+                .collect(),
         }
     }
 }
@@ -166,11 +183,10 @@ impl TryFrom<IngressPayload> for Vec<SignedIngress> {
     type Error = IngressPayloadError;
     fn try_from(payload: IngressPayload) -> Result<Vec<SignedIngress>, Self::Error> {
         payload
-            .id_and_pos
-            .iter()
-            .enumerate()
-            .map(|(i, _)| payload.get(i).map(|m| m.1))
-            .collect::<Result<_, _>>()
+            .ingress_messages
+            .into_iter()
+            .map(SignedIngress::try_from)
+            .collect()
     }
 }