--- a/rs/types/types/src/exhaustive.rs
+++ b/rs/types/types/src/exhaustive.rs
@@ -1,5 +1,6 @@
 //! Implementations and serialization tests of the ExhaustiveSet trait
 
+use crate::messages::{Blob, Delegation, HttpCallContent, HttpCanisterUpdate, HttpRequestEnvelope, SignedDelegation, SignedIngress};
 use crate::consensus::hashed::Hashed;
 use crate::consensus::idkg::common::{PreSignatureInCreation, PreSignatureRef};
 use crate::consensus::idkg::ecdsa::{QuadrupleInCreation, ThresholdEcdsaSigInputsRef};
@@ -31,6 +32,7 @@ use crate::signature::{
     BasicSignature, BasicSignatureBatch, MultiSignature, MultiSignatureShare, ThresholdSignature,
     ThresholdSignatureShare,
 };
+use crate::time::UNIX_EPOCH;
 use crate::xnet::CertifiedStreamSlice;
 use crate::{CryptoHashOfState, ReplicaVersion};
 use ic_base_types::{CanisterId, NodeId, PrincipalId, RegistryVersion, SubnetId};
@@ -904,6 +906,51 @@ fn replace_by_singleton_if_empty<
     });
 
     BTreeMap::from([(key, value)])
+}
+
+impl ExhaustiveSet for SignedIngress {
+    fn exhaustive_set<R: RngCore + CryptoRng>(rng: &mut R) -> Vec<Self> {
+		let ingress_expiry = UNIX_EPOCH;
+        let content = HttpCallContent::Call {
+            update: HttpCanisterUpdate {
+                canister_id: Blob(vec![42; 8]),
+                method_name: "some_method".to_string(),
+                arg: Blob(b"".to_vec()),
+                sender: Blob(vec![0x05]),
+                nonce: Some(Blob(vec![1, 2, 3, 4])),
+                ingress_expiry: ingress_expiry.as_nanos_since_unix_epoch(),
+            },
+        };
+        let update_messages = vec![
+            HttpRequestEnvelope::<HttpCallContent> {
+                content: content.clone(),
+                sender_pubkey: Some(Blob(vec![2; 32])),
+                sender_sig: Some(Blob(vec![1; 32])),
+                sender_delegation: None,
+            },
+            HttpRequestEnvelope::<HttpCallContent> {
+                content: content.clone(),
+                sender_pubkey: None,
+                sender_sig: None,
+                sender_delegation: None,
+            },
+            HttpRequestEnvelope::<HttpCallContent> {
+                content,
+                sender_pubkey: Some(Blob(vec![2; 32])),
+                sender_sig: Some(Blob(vec![1; 32])),
+                sender_delegation: Some(vec![SignedDelegation::new(
+                    Delegation::new(vec![1, 2], ingress_expiry),
+                    vec![3, 4],
+                )]),
+            },
+        ];
+        let signed_ingresses: Vec<SignedIngress> = update_messages
+            .into_iter()
+            .map(|msg| SignedIngress::try_from(msg).unwrap())
+            .collect();
+
+		vec![]
+    }
 }
 
 /// Some maps of the form Map<Id, (Id, Data)> are serialized as vectors of their values, throwing