The current implementation delays the deserialization of the ingress messages until they are accessed.
An upside of this implementation is that if we never access the ingress messages - we will never deserialize them.
An downside is that it's quite complex, error prone, and hard to modify. Moreover, when we actually need to access *all* ingress messages in the payload, the simpler implementation is actually more efficient (at least according to the benchmarks below)

# Benchmarks
Ran `load_blocks` benchmark via `bazel run //rs/artifact_pool:load_blocks_bench` with a modification that each block has 100 ingress messages of 128KB each. (Note that in the benchmark we load all blocks in the pool, i.e. 20)

|  |  Load blocks and sum their heights | Load blocks and sum their ingress counts | Load blocks and sum their ingress nonces|
| - | - | - | - |
| old struct times | 39.260 µs | 25.151 ms | 714.39 ms |
| new struct times |  39.978 µs |  80.508 ms | 173.15 ms |
| change | +2.2538% | +220.09% | -75.762% |

## Conclusions
1. Loading blocks without their payloads is as fast as before (not surprisingly)
2. Loading blocks *and* their payloads without accessing ingresses messages is 3x slower
3. Loading blocks *and* their payload *and* accessing all the ingress messages is 4x faster

# Backwards compatibility
The changes are not backwards compatible but it's okay as we don't preserve `IngressPayload`s across replica versions

# Tests
Ran a benchmark where we send 4 * 1MB ingress messages per second to a subnet with 13 nodes and looked at the flamegraph:
![flamegraph_big_messages](https://github.com/user-attachments/assets/5f1a58e0-0666-45ff-aa4e-8198baa6c7c1)
I didn't observe any deterioration